Changed the order of later filtration (e.g., invokerType, argumentTypes)
Previously, later filterations were done in another method from the
output of initial matching (which is done in 4 steps). So if any one
step, a match is found we return to the method of filtration. Now if
we do not found eligible method after further filtration there was no
way to go back to next steps. Moved the filtration call inside the
method that does initial matching so that we can visit next steps in
case we do not find any eligible method after extra filtration.