Improved exact matching criteria for method argument matching
According to previous implementation, we carry an boolean indicator
of exact match of method argument match if all argument exactly matches
with the provided onces. Now we carry the number of exact match
arguments for that method. At last, we return the method which has
maximum number of matched arguments.

GH-8