Copied invoker's type arguments if class info of method is super class
If class info of method is super class of invoker we previously
assigned the class info as invoker to get the appropriate formal type
parameters. Now we have also copied the type arguments from invoker to
class info of method. So that we can resolve type arguments from invoker
properly in this case.