Added option to pass storage file name
In order to reduce size of storage file, added option to
pass storage file name. Also, added mechanism to clear
caching. So we are sacrificing time efficiency to solve
the out of memory issue.

Until we move to a storage based option, we can use this feature
to solve the memory issue.